+++ crates/atuin-client/src/encryption.rs
@@ -10,7 +10,7 @@
 
 use std::io::prelude::*;
 
-pub use crate::record::encryption::PasetoV4Key;
+pub use atuin_common::encryption::paseto_v4;
 use base64::prelude::{BASE64_STANDARD, Engine};
 use crypto_secretbox::{KeyInit, XSalsa20Poly1305, aead::OsRng};
 use eyre::{Context, Result, bail, ensure, eyre};
@@ -19,14 +19,14 @@ use rmp::Marker;
 
 use crate::settings::Settings;
 
-pub fn generate_encoded_key() -> Result<(PasetoV4Key, String)> {
-    let key = PasetoV4Key::from(<[u8; 32]>::from(XSalsa20Poly1305::generate_key(&mut OsRng)));
+pub fn generate_encoded_key() -> Result<(paseto_v4::Key, String)> {
+    let key = paseto_v4::Key::from(<[u8; 32]>::from(XSalsa20Poly1305::generate_key(&mut OsRng)));
     let encoded = encode_key(&key)?;
 
     Ok((key, encoded))
 }
 
-pub fn new_key(settings: &Settings) -> Result<PasetoV4Key> {
+pub fn new_key(settings: &Settings) -> Result<paseto_v4::Key> {
     let path = &settings.key_path;
 
     if path.exists() {
@@ -42,7 +42,7 @@ pub fn new_key(settings: &Settings) -> Result<PasetoV4Key> {
 }
 
 // Loads the secret key, will create + save if it doesn't exist
-pub fn load_key(settings: &Settings) -> Result<PasetoV4Key> {
+pub fn load_key(settings: &Settings) -> Result<paseto_v4::Key> {
     let path = &settings.key_path;
 
     let key = if path.exists() {
@@ -55,7 +55,7 @@ pub fn load_key(settings: &Settings) -> Result<PasetoV4Key> {
     Ok(key)
 }
 
-pub fn encode_key(key: &PasetoV4Key) -> Result<String> {
+pub fn encode_key(key: &paseto_v4::Key) -> Result<String> {
     let key = key.as_bytes();
     let mut buf = vec![];
     rmp::encode::write_array_len(&mut buf, key.len() as u32)
@@ -69,7 +69,7 @@ pub fn encode_key(key: &PasetoV4Key) -> Result<String> {
     Ok(buf)
 }
 
-pub fn decode_key(key: String) -> Result<PasetoV4Key> {
+pub fn decode_key(key: String) -> Result<paseto_v4::Key> {
     use rmp::decode;
 
     let buf = BASE64_STANDARD
@@ -115,7 +115,7 @@ pub fn decode_key(key: String) -> Result<PasetoV4Key> {
 mod test {
     #[test]
     fn key_encodings() {
-        use super::{PasetoV4Key, decode_key, encode_key};
+        use super::*;
 
         // a history of our key encodings.
         // v11.0.0 xCAbWypb0msJ2Kq+8j4GVEWUlDX7deKnrTRSIopuqXxc5Q==
@@ -130,7 +130,7 @@ mod test {
         // b8b57c8 xCAbWypb0msJ2Kq+8j4GVEWUlDX7deKnrTRSIopuqXxc5Q==                     (https://github.com/atuinsh/atuin/pull/1057)
         // 8c94d79 3AAgG1sqW8zSawnM2MyqzL7M8j4GVEXMlMyUNcz7dczizKfMrTRSIsyKbsypfFzM5Q== (https://github.com/atuinsh/atuin/pull/1089)
 
-        let key = PasetoV4Key::from([
+        let key = paseto_v4::Key::from([
             27, 91, 42, 91, 210, 107, 9, 216, 170, 190, 242, 62, 6, 84, 69, 148, 148, 53, 251, 117,
             226, 167, 173, 52, 82, 34, 138, 110, 169, 124, 92, 229,
         ]);
