+++ crates/atuin-client/src/record/mod.rs
@@ -1,4 +1,3 @@
-pub mod encryption;
 pub mod sqlite_store;
 
 #[cfg(feature = "sync")]
